pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "milo"
version = "0.1.0"
description = "Template-driven CLI applications for free-threaded Python"
requires-python = ">=3.14"
license = "MIT"
dependencies = ["kida-templates>=0.2.9"]
keywords = ["cli", "terminal", "forms", "free-threading", "template", "elm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.14",
    "Topic :: Terminals",
    "Typing :: Typed",
]

[project.scripts]
milo = "milo.cli:main"

[project.optional-dependencies]
html = []

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
milo = ["templates/*.txt", "py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["milo"]
branch = true

[tool.coverage.report]
fail_under = 80

[tool.ty.environment]
python-version = "3.14"

[tool.ty.rules]
call-non-callable = "warn"
unresolved-attribute = "warn"

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "B", "SIM", "I", "N", "PIE", "PERF", "C4", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101", "SIM117", "N806"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.100.0",
    "ty>=0.0.11",
    "ruff>=0.15.1",
]
